+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Articles are now sorted by publish date
 - **Document Changes**
   - Featured documents are now display on the Documents Overview
+  - Documents can now be renamed
 - **Gem Changes**
   - Update to ruby 2.7.1
   - Update to rails 6.0.3.3